User commit

+++ src/main/java/com/example/demo/service/ValidationService.java
@@ -3,5 +3,5 @@ import com.example.demo.entity.ValidationEntity;
 import java.util.*;
 public interface ValidationService{
     ValidationEntity post(ValidationEntity stu);
-    ValidationEntity finddata(int id);
+    ValidationEntity finddata(Long id);
 }
